re-named NotesContainer.js to ProgressNotes.js broke out form in ProgressView into its own component called ProgressForm.js moved ProgressView into containers folder
