+++ front-end/src/App.js
@@ -1,22 +1,22 @@
-import React, { Component } from "react";
-import { Route } from "react-router-dom";
+import React, { Component } from 'react';
+import { Route } from 'react-router-dom';
 
-import auth from "./Auth";
-import "./components/styles/App.sass";
-import NavBar from "./components/NavBar";
-import LandingPage from "./components/LandingPage";
-import Callback from "./Callback";
-import ScheduleView from "./components/ScheduleView";
-import WorkoutsView from "./containers/WorkoutsView";
-import ProgressView from "./components/ProgressView";
-import BillingView from "./components/BillingView";
-import SettingsView from "./components/SettingsView";
-import SecuredRoute from "./components/SecuredRoute";
+import auth from './Auth';
+import './components/styles/App.sass';
+import NavBar from './components/NavBar';
+import LandingPage from './components/LandingPage';
+import Callback from './Callback';
+import ScheduleView from './components/ScheduleView';
+import WorkoutsView from './containers/WorkoutsView';
+import ProgressView from './containers/ProgressView';
+import BillingView from './components/BillingView';
+import SettingsView from './components/SettingsView';
+import SecuredRoute from './components/SecuredRoute';
 
 class App extends Component {
   componentDidMount() {
     const { renewSession } = auth;
-    if (localStorage.getItem("isLoggedIn") === "true") {
+    if (localStorage.getItem('isLoggedIn') === 'true') {
       console.log(auth);
       renewSession();
     }
