+++ front-end/src/components/styles/ProgressNotes.sass
